# cortes-ia-contracts/design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortes IA — contrato proposto",
    "version": "0.2.0",
    "description": "Contrato-alvo atualizado pelas decisões v0.2. Não confundir com OpenAPI efetivamente gerado pela API inicial; ver API_IMPLEMENTADA.md."
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Placeholder deliberado; substituir pelo host do ambiente."
    }
  ],
  "paths": {
...
    "/api/v1/catalog": {
      "get": {
        "operationId": "catalog_get",
        "summary": "Consultar catálogo publicado e limites",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "catalog"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet": {
      "get": {
        "operationId": "wallet_get",
        "summary": "Consultar saldo discriminado",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/transactions": {
      "get": {
        "operationId": "wallet_transactions",
        "summary": "Consultar extrato",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
...
    "/api/v1/projects/uploads": {
      "post": {
        "operationId": "projects_upload",
        "summary": "Aceitar upload privado e criar histórico",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCreated"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{uploadId}/complete": {
      "post": {
        "operationId": "projects_completeUpload",
        "summary": "Concluir bytes e agendar validação",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Operação aceita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUpload"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/imports": {
      "post": {
        "operationId": "projects_import",
        "summary": "Aceitar link para preflight seguro",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Operação aceita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "projects_list",
        "summary": "Consultar histórico",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/api/v1/projects/{projectId}": {
      "get": {
        "operationId": "projects_get",
        "summary": "Consultar projeto",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Versão para If-Match."
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "delete": {
        "operationId": "projects_delete",
        "summary": "Marcar exclusão e remover mídia",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Operação aceita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/quotes": {
      "post": {
        "operationId": "projects_quote",
        "summary": "Calcular custo autoritativo",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/runs": {
      "post": {
        "operationId": "projects_run",
        "summary": "Confirmar custo e reservar execução",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Operação aceita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cortes_session"
      },
      "serviceAuth": {
        "type": "mutualTLS",
        "description": "Workload autenticado no ingresso privado; sem acesso público."
      }
    },
    "schemas": {
...
      "Catalog": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tier"
            }
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            }
          },
          "combos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Combo"
            }
          },
          "packages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Package"
            }
          },
          "presets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Preset"
            }
          },
          "maxUploadBytes": {
            "type": "integer",
            "default": 5000000000,
            "description": "Limite inicial aprovado, ajustável por política administrativa."
          },
          "maxDurationMs": {
            "type": "integer",
            "default": 10800000
          },
          "maxClips": {
            "type": "integer",
            "minimum": 1
          },
          "supportedSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptedMediaTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "version",
          "tiers",
          "features",
          "combos",
          "packages",
          "presets",
          "maxUploadBytes",
          "maxDurationMs",
          "maxClips",
          "supportedSources",
          "acceptedMediaTypes"
        ]
      },
...
    }
  }
}